LibJS: Narrow bytecode source map offsets from size_t to u32
Add VERIFY guards to catch bytecode programs that exceed u32::max bytes
and narrow the bytecode_offset parameter in add_source_map_entry() to
u32. This is a preparatory change for optimizing source map storage.

+++ Libraries/LibJS/Bytecode/BasicBlock.h
@@ -56,7 +56,7 @@ public:
     BasicBlock const* finalizer() const { return m_finalizer; }
 
     auto const& source_map() const { return m_source_map; }
-    void add_source_map_entry(size_t bytecode_offset, SourceRecord const& source_record) { m_source_map.set(bytecode_offset, source_record); }
+    void add_source_map_entry(u32 bytecode_offset, SourceRecord const& source_record) { m_source_map.set(bytecode_offset, source_record); }
 
     [[nodiscard]] bool has_resolved_this() const { return m_has_resolved_this; }
     void set_has_resolved_this() { m_has_resolved_this = true; }
